changed quantity data type bug

+++ invoices/models.py
@@ -291,10 +291,7 @@ class InvoiceItem(models.Model):
 
     hsn_sac = models.CharField(max_length=100)
 
-    quantity = models.DecimalField(
-        max_digits=10,
-        decimal_places=2
-    )
+    quantity = models.PositiveIntegerField()
 
     rate = models.DecimalField(
         max_digits=10,
